bswap.h: use built-in bswap32 and bswap64 when available
Clang 12 and GCC 10 added support for built-in bswap32 and bswap64. As bswap.h doesn't support ARM64/aarch64 when using GCC/Clang, this is a good opportunity to start using those built-ins. In the future, we might want to remove the custom git_bswap32 and git_bswap64 altogether and leave that up to the compiler to handle.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>