CMakeLists.txt: enable building Git using LLVM/Clang on Windows
While Clang is available on Windows, it uses the Visual Studio linker by default:

"For Windows projects, Visual Studio by default invokes Clang in clang-cl mode. It links with the Microsoft implementation of the Standard Library."

https://docs.microsoft.com/en-us/cpp/build/clang-support-msbuild?view=msvc-160

Since we want to build with the Clang/llvm-mingw toolchain, we have to get rid of some assumptions in the CMakeLists.txt file, as it's basically a GNU compiler and linker.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>